move TransposeNormal<platform::CPUDeviceContext, T> to .cc file
