+++ paddle/fluid/operators/math/math_function_impl.h
@@ -13,11 +13,9 @@ See the License for the specific language governing permissions and
 limitations under the License. */
 
 #pragma once
-#include <utility>
 #include <vector>
 #include "paddle/fluid/framework/data_type.h"
 #include "paddle/fluid/operators/math/math_function.h"
-#include "unsupported/Eigen/CXX11/Tensor"
 
 namespace paddle {
 namespace operators {
@@ -32,40 +30,6 @@ void SetConstant<DeviceContext, T>::operator()(const DeviceContext& context,
   auto t = framework::EigenVector<T>::Flatten(*tensor);
   t.device(*context.eigen_device()) = t.constant(static_cast<T>(num));
 }
-
-template <typename T>
-struct TransposeNormal<platform::CPUDeviceContext, T> {
-  void operator()(const platform::CPUDeviceContext& context,
-                  const framework::Tensor& in, framework::Tensor* out,
-                  const std::vector<int>& axis) {
-    const int rank = axis.size();
-    auto in_stride = framework::stride(in.dims());
-    auto out_stride = framework::stride(out->dims());
-    const T* in_ptr = reinterpret_cast<const T*>(in.data<T>());
-    T* out_ptr = reinterpret_cast<T*>(out->data<T>());
-
-    auto transpose_helper = [&](int64_t beg, int64_t end) {
-      for (int64_t out_idx = beg; out_idx < end; ++out_idx) {
-        int64_t in_idx = 0;
-        int64_t tmp_idx = out_idx;
-        // calculate the input index
-        for (int i = 0; i < rank; ++i) {
-          const int64_t coordinate = tmp_idx / out_stride[i];
-          tmp_idx -= coordinate * out_stride[i];
-          in_idx += coordinate * in_stride[axis[i]];
-        }
-        out_ptr[out_idx] = in_ptr[in_idx];
-      }
-    };
-    double cost_per_iteration =
-        rank * (Eigen::TensorOpCost::DivCost<int64_t>() +
-                2 * Eigen::TensorOpCost::MulCost<int64_t>() +
-                2 * Eigen::TensorOpCost::AddCost<int64_t>());
-    Eigen::TensorOpCost cost(sizeof(T), sizeof(T), cost_per_iteration);
-    auto* cpu_device = context.eigen_pool_device();
-    cpu_device->parallelFor(out->numel(), cost, std::move(transpose_helper));
-  }
-};
 
 template <typename DeviceContext, typename T, int Rank>
 void Transpose<DeviceContext, T, Rank>::operator()(
